Update search -> userList component
