+++ apps/client/src/components/index.ts
@@ -3,3 +3,4 @@ export { default as Header } from './header';
 export { default as AlarmModal } from './alarmModal';
 export { default as WriteButton } from './writeButton';
 export { default as UserItem } from './userItem';
+export { default as UserList } from './userList';
