+++ apps/client/src/stores/index.ts
@@ -1,0 +1,1 @@
+export { default as useSelectedIdState } from './useSelectedIdStore';
